requirements.txt
# Django and Web Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7
django-extensions==3.2.3

# Data Analysis
pandas==2.1.3
numpy==1.24.3
networkx==3.2.1

# File Processing
Pillow==10.1.0
python-magic==0.4.27

# API Documentation
drf-yasg==1.21.7

# Development Tools
python-decouple==3.8
whitenoise==6.6.0

# Geographic Data (Optional - for future PostGIS integration)
# django-geojson==4.0.0
# django-leaflet==0.28.3

# Testing
factory-boy==3.3.0
faker==20.1.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
